sandbox: Refactor code to create os_jump_to_file()
At present os_jump_to_image() jumps to a given image, and this is written
to a file. But it is useful to be able to jump to a file also.

To avoid duplicating code, split out the implementation of
os_jump_to_image() into a new function that jumps to a file.

Change-Id: I0feb5b7a5faa4ea0df30c87b572586723dd55a11
Signed-off-by: Simon Glass <sjg@chromium.org>