test: Add a 'make qcheck' target for quicker testing
At present tests are quite slow to run, over a minute on my machine. This
presents a considerable barrier to bisecting for failures.

The slowest tests are the filesystem ones and the buildman --fetch-arch
test. Add a new 'qcheck' target that skips these tests. This reduces test
time down to about 40 second, still too long, but bearable.

Change-Id: I69185dbb3cbf1482db02fda4e2ae307aab7b97b5
Signed-off-by: Simon Glass <sjg@chromium.org>