RFC: Makefile: Build U-Boot as a library
As an experiment, build U-Boot as a library, u-boot.o, so it can be used
by other open-source software. Update the sandbox build rules to suit.

This has not been tested in any meaningful way. It breaks sandbox_spl and
all x86 boards.

I am interested in feedback as to how useful this might be.

Signed-off-by: Simon Glass <sjg@chromium.org>
Series-to: u-boot
Cover-letter:
Various patches for verified boot support
This is (for now) the last series to enable Chromium OS verified boot in
U-Boot. It includes:

- Sandbox valgrind spport
- Minor TPM enhancements and fixes
- Resurection of inttypes.h
- Args checking for log functions
- Sandbox fix for time jitter
- Various other minir things
END