Add an empty stdint.h file
Some libraries build by U-Boot may include stdint.h. This is not used by
U-Boot itself and causes conflicts with the types defined in
linux/types.h. To work around this, add an empty file with this name so
that it will be used in preference to the compiler version.

Change-Id: Ieecb387277c537d30c6521fce9ce9f129a6e053f
Signed-off-by: Simon Glass <sjg@chromium.org>
Series-to: u-boot
Series-cc: masa