gpio: Show inactive GPIOs when explicitly requested
At present the gpio command only shows GPIOs which are marked as in use.
This makes sense with 'gpio status' since we already have the '-a' flag
to indicate that all GPIOs should be shown. But when a particular GPIO is
requested, it seems better to always display it. At present the request is
simply ignored.

For example if GPIO a10 is not in use, then:

   > gpio a10

shows nothing, not even the function being used for that GPIO. With this
change, it shows the pin status:

   > gpio a10
   a10: input: 0 [ ]

Add an extra parameter for this to avoid changing the existing flag
parameter.

Series-changes: 2
- Add an example to the commit message for clarity

Signed-off-by: Simon Glass <sjg@chromium.org>