x86: broadwell: Don't bother probing the PCH for pinctrl
At present the pinctrl probes the PCH but since it only uses it to obtain
a PCI address, this is no necessary. Avoiding this fixes one of the two
co-dependent loops in broadwell.

This driver really should be a proper pinctrl driver, but for now it
remains a syscon device.

Change-Id: I17366834782863d6da578c659375d5dbbc777dac
Signed-off-by: Simon Glass <sjg@chromium.org>