x86: broadwell: Allow SDRAM init from SPL
At present SDRAM is always set up in U-Boot proper. Allow this to be done
in SPL instead so that U-Boot proper can be loaded into SDRAM and run
from there. This allows U-Boot to be compressed to reduce space, since
it is not necessary to run it directly from flash.

Signed-off-by: Simon Glass <sjg@chromium.org>