x86: spi: Add helper functions for Intel Fast SPI
Most x86 CPUs use a mechanism where the SPI flash is mapped into the very
top of 32-bit address space, so that it can be executed in place and read
simply by copying from memory. For an 8MB ROM the mapping starts at
0xff800000.

However some recent Intel CPUs do not use a simple 1:1 memory map. Instead
the map starts at a different address and not all of the SPI flash is
accessible through the map. This 'Fast SPI' feature requires that U-Boot
check the location of the map. It is also possible (optionally) to read
from the SPI flash using a driver.

Add support for booting from Fast SPI. The memory-mapped version is used
by both TPL and SPL on Apollo Lake.

In respect of a SPI flash driver, the actual SPI driver is ich.c - this
just adds a few helper functions and definitions.

This is used by Apollo Lake.

Series-changes: 3
- Add the missing header file
- Add support for of-platdata for TPL
- Use pci_get_devfn()
- Reorder file so that write() and erase() are together
- Don't include write() and erase() in TPL
- Change Fast-SPI driver into a helper file used by ICH SPI
- Merge in patch "x86: Add support for booting from Fast SPI"

Series-changes: 3
- Drop 'a4' comment for register offset

Change-Id: I24ab1b373d29b4c5bf381f2e3a668c802e316a32
Signed-off-by: Simon Glass <sjg@chromium.org>