common: Move serial_printf() to the serial header
Move this function header to serial.h since this function is clearly
related to serial. The function itself stays in console.c since we don't
have a single serial file, DM and non-DM each has a separate file so we
would have to either create a new common serial file, or repeat the
function in both serial.c and serial-uclass.c, neither of which seem
worthwhile.

Change-Id: Ia99e87eadc32133a56d8c55dbcca4a687743fdd8
Signed-off-by: Simon Glass <sjg@chromium.org>