dm: pci: Move pci_get_devfn() into a common file
Early in boot it is necessary to decode the PCI device/function values for
particular peripherals in the device tree or of-platdata. This is needed
in TPL where CONFIG_PCI is not defined.

To handle this, move pci_get_devfn() into a file that is built even when
CONFIG_PCI is not defined.

Also add a function for use by of-platdata, to convert a reg property to
a pci_dev_t.

Series-changes: 3
- Update device type to pci_dev_t
- Move the function to a common file instead of duplicating it

Series-changes: 4
- Rename pci_x86_ofplat_get_devfn() to pci_ofplat_get_devfn()
- Add more documentation for pci_ofplat_get_devfn()
- Mention that the return value is pci_dev_t

Signed-off-by: Simon Glass <sjg@chromium.org>