x86: Update mrccache to support multiple caches
With Apollo Lake we need to support a normal cache, which almost never
changes and a much smaller 'variable' cache which changes every time.

Update the code to add a cache type, use an array for the caches and use a
for loop to iterate over the caches.

Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Bin Meng <bmeng.cn@gmail.com>
Series-changes: 3
- Move line related to variable-cache into the next patch

Series-changes: 4
- apollolake -> Apollo Lake

Change-Id: Ie8badf1328570408b3352525d35cd0f804073e17