x86: Add chromebook_coral
Add support for coral which is a range of apollolake-based Chromebook
released in 2017. This also includes reef released in 2016, since it is
based on the same SoC.

Series-to: u-boot
Series-cc: bin
Cover-letter:
x86: Add initial support for apollolake
Apollolake is an Intel SoC generation aimed at relatively low-end embedded
systems. It was released in 2016 but has become more popular recently with
some embedded boards using it.

This series adds support for apollolake. As an example it adds an
implementation of chromebook_coral (a large range of Chromebooks released
in 2017).

The series provides only enough to boot to a prompt. It does not include
display, storage, USB, etc. But it does include the major part of the
plumbing.

Since this is the first time U-Boot has used FSP2 there is quite a bit of
refactoring needed. Some of this is in previous series, but much of it is
here.

Some further clean-up is needed with this series and it needs more
documentation, but it is ready for review.

This series is available at u-boot-dm/coral-working
END

Change-Id: Ied9423f679557f95a4714a72bd7aa11fdd4445ac
Signed-off-by: Simon Glass <sjg@chromium.org>