x86: Add an option to control the position of U-Boot
The existing work-around for positioning U-Boot in the ROM when it
actually runs from RAM still exists and there is not obvious way to change
this.

Add a proper Kconfig option to handle this case. This also adds a new bool
property to indicate whether CONFIG_SYS_TEXT_BASE exists.

Series-changes: 4
- Rename option to HAVE_SYS_TEXT_BASE

Change-Id: I453a662d790c8ee5e3ec1a27c8e29feb84d03087
Signed-off-by: Simon Glass <sjg@chromium.org>