tpm: Add a driver for H1/Cr50
H1 is a Google security chip present in recent Chromebooks, Pixel phones
and other devices. Cr50 is the name of the software that runs on H1 in
Chromebooks.

This chip is used to handle TPM-like functionality and also has quite a
few additional features.

Add a driver for this.

Series-changes: 2
- Significant rewrite of cr50 init procedure
- Support use of interrupts

Signed-off-by: Simon Glass <sjg@chromium.org>