board_r: Move early-timer init later
At present the early timer init happens as soon as driver mode is set up.
This makes it impossible to do any in that needs driver model but must run
before devices are problem (as needed with Intel's FSP-S, for example).

In any case it is not a good idea to tie probing of particular drivers too
closely to the DM init.

Create a new function to init the timer and put it a bit later in the
sequence.

Signed-off-by: Simon Glass <sjg@chromium.org>
Series-changes: 4
- Add new patch to move early-timer init later