Add the nuklear library
U-Boot supports based display and keyboard-input functionaliy which can be
used to show boot menus and the like. However this is fairly rudimentary
and is not really suitable for producing a user-facing UI.

With verified boot or firmware-based diagnostics, users expect some sort
of friendly interface, perhaps similar to that provided by the UEFI BIOS,
for example.

The Nuklear library provides an easy way to build GUI into U-Boot. Add
the source code from commit 1de7702.

Signed-off-by: Simon Glass <sjg@chromium.org>