dm: core: Add a new flag to track platform data
We want to avoid allocating platform data twice. This could happen if
device_probe() is called after device_ofdata_to_platdata() for the same
device.

Add a flag to track whether device_ofdata_to_platdata() has been called on
a device. Check the flag to make sure it doesn't happen twice, and clear
the flag when the data is freed.

Change-Id: Iaee21ff41733fc4ee01284e3a7ca6f57874a1a1b
Signed-off-by: Simon Glass <sjg@chromium.org>