i2c: Add enums for i2c speed and address size
Some drivers define their own speed enums and use their own constants for
speed. It makes sense to have a unified defition of the different speeds.

Since many controllers have to do different things for fast/high speed, it
is a good idea to have an enum for the mode.

Add these as well as an enum for the address mode.

Change-Id: Ie472f8a2d709a2cb8383826b213525abd5d5598a
Signed-off-by: Simon Glass <sjg@chromium.org>