dm: core: Require users of devres to include the header
At present devres.h is included in all files that include dm.h but few
make use of it. Also this pulls in linux/compat which adds several more
headers. Drop the automatic inclusion and require files to include devres
themselves. This provides a good indication of which files use devres.

Change-Id: I451105ed1149334d1ef6eca303eede6e56b833d1
Signed-off-by: Simon Glass <sjg@chromium.org>