console: Add a function to read a line of the output / eof
When recording the console output for testing it is useful to be able to
read the output a line at a time to check that the output is correct. Also
we need to check that we get to the end of the output.

Add a console function to return the next line and another to see how must
data is left.

Change-Id: I8114ed3d67267c833b7d9f456e6da6e9c4653c9b
Signed-off-by: Simon Glass <sjg@chromium.org>