test: Add a way to check each line of console output
When writing tests to check the output from commands it is useful to be
able to check the output line by line using an assertion. Add a helper
macro to support this and to check that there is no unexpected trailing
data.

Also some commands product a dump using print_buffer(). Add a way to check
that the correct number of bytes are dumped (ignoring the actual
contents).

Change-Id: I6e63c1c6aae518ed5a245427b8046b1bac3dbf66
Signed-off-by: Simon Glass <sjg@chromium.org>