acpi: Record the items added to SSDT
It is useful to be able to control the order of data written to the SSDT
so that we can compare the output against known-good kernel dumps.

Add code to record each item that is added along with the device that
added it. That allows us to reorder things later if needed.

Change-Id: I1f119e7a847da3e0b2b1791e6f48acaeaf25a223
Signed-off-by: Simon Glass <sjg@chromium.org>