x86: acpi: Support external GNVS tables
At present U-Boot puts a magic number in the ASL for the GNVS table and
searches for later.

Add a Kconfig option to use a different approach, where the ASL files
declare the table as an external symbol. U-Boot can then put it wherever
it likes, without any magic numbers or searching.

Change-Id: Ic4e4767888256deed0503035fd4cc5c4d1060a62
Signed-off-by: Simon Glass <sjg@chromium.org>