Add a few floating-point routines
U-Boot does not normally make use of floating point and it is
generally possible and desirable to avoid it within a bootloader.

However there are libraries which need it. For example, U-Boot's
Truetype font implementation needs floating point and some features
such as the Nuklear GUI make use of it also.

Enable this to compile in some basic library functions for floating
point.

Signed-off-by: Simon Glass <sjg@chromium.org>