video: Add support for simple interactive menus/features
U-Boot provides display support but currently lacks the ability to show
simple graphical user interfaces to users, like UEFI, for example. It
would be very useful to be able to display a nice boot menu, for example.

Add a new GUI uclass which handles display and user input. This can be
used to deal with the basics of such features.

Signed-off-by: Simon Glass <sjg@chromium.org>