test/py: Allow using buildman to build U-Boot
It is a pain to have to set the CROSS_COMPILE environment variable when
using test.py's --build option. It is possible to get this using the -A
option from buildman. But it seems better to just use buildman to do the
build when it is available.

However using buildman adds a new dependency to the test system which we
want to avoid. So leave the default as is and add a flag to make it use
buildman.

Note that most of these changes relate to test.py and the parts of the
travis/gitlab/azure scripts which relate to running test and building a
suitable U-Boot to run the tests on.

Signed-off-by: Simon Glass <sjg@chromium.org>
Series-changes: 2
- Convert this into an option, leaving the default as is

Series-to: u-boot
Series-cc: stephen, trini, bin
Series-process-log: sort, uniq
Series-version: 3
Cover-letter:
gitlab: Simplify the test script
At present there are several things in the gitlab, travis and azure
scripts which work around limitations in buildman. With a few small
feature additions these can be removed.

This series adds some new features to buildman and simplifies the script:
- Option to run a single build in a specified output directory
- Allow ignoring warnings
- Removes a restriction on the build output directory

It also
- adds a way for test.py to use buildman for the --build option
- makes one change to azure since the same approach should be possible there
- fixes a few minor problems noticed in main and sandbox docs
- removes the -a buildman flag which serves no purpose

Note: I tested that this works OK on travis, gitlab and azure:
https://travis-ci.org/github/sglass68/u-boot/builds/662740008
https://gitlab.denx.de/u-boot/custodians/u-boot-dm/pipelines/2435
https://dev.azure.com/simon0972/u-boot/_build/results?buildId=3&view=results

There is something funny about azure - the test called
"Run binman, buildman, dtoc and patman testsuites" fails the 'buildman -t'
test but azure reports success.
END