rockchip: Drop the fit_spl_optee.sh script
Now that all board use binman instead of this script, drop it.

Series-to: u-boot
Series-cc: rockchip, Jagan Teki <jagan@amarulasolutions.com>
Series-version: 2
Series-changes: 2
- Add patches to partially migrate rockchip to use binman

Cover-letter:
rockchip: Support building ROM files automatically with binman
Rockchip-based Chromebooks support booting from SPI flash. It is annoying
to have to manually build the SPI image when the SD image is built
automatically.

This feature is already available for x86 devices, so the existing
mechanism is reused. Briefly, this allows a BUILD_ROM environment variable
to be provided to indicate that any required binary blobs are present and
it is safe to build the ROM.

A new 'mkimage' type is added to binman to support building binaries
containing mkimagem using a binman definition to configure it. This avoids
Makefile/shell/Python code to do the same thing.

This series also migrates some rockchip boards to use binman to produce
their FIT as well, resulting in removing the fit_spl_optee.sh script.

Other archs and the rest of rockchip could be migrated too.

This series uses binman to produce a ROM image on two selected
Chromebooks, Bob (RK3399) and Jerry (RK3388).
END

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I16748de85e34ae0361fae59525e3c8c0bb8fe734