FROMLIST: binman: Move GetEntryModules() to control
When binman is installed its main program is in a different directory
to its modules. This means that __file__ is different and we cannot use
it to obtain the path to etype/ from main.py

To fix this, move the function to the 'control' module, since it is
installed with all the other modules, including the etype/ directory.

BUG=chromium:1112434
TEST=binman test

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I63ed3a0183d8a14aba024a6cf7c27c1503f51ef7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/u-boot/+/2354363
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>