smbios: Add more properties
The current tables only support a subset of the available fields defined
by the SMBIOS spec. Add a few more.

We could use CONFIG_SYS_CPU or CONFIG_SYS_SOC as a default for family, but
the meaning of that value relates more to the whole system rather than
just the SoC.

Series-change: 5
- Update commit message to explain why family does not have a default

Signed-off-by: Simon Glass <sjg@chromium.org>