dm: Support parent devices with of-platdata
At present of-platdata does not provide parent information. But this is
useful for I2C devices, for example, since it allows them to determine
which bus they are on.

Add support for setting the parent correctly, by storing the parent
driver_info index in dtoc and reading this in lists_bind_drivers(). This
needs multiple passes since we must process children after their parents
already have been bound.

Series-changes: 2
- Make sure devices whose parents fail to bind are still themselves bound
- Add a few more comments
- Add a special case for !OF_PLATDATA_PARENT

Signed-off-by: Simon Glass <sjg@chromium.org>