checkpatch: Add warnings for unexpected struct names
As a way of keeping the driver declarations more consistent, add a warning
if the struct used does not end with _priv or _plat.

Series-to: u-boot
Cover-letter:
dm: treewide: Tidy up and adjust driver model naming
There are several naming problems in driver model which I think are worth
sorting out.

FIrstly, the _auto_alloc_size suffix is widely used but is quite
long-winded. In earlier days it made some sense since auto allocation was
a strange concept. But with driver model now used pretty universally, we
can shorten this to 'auto'. This reduces verbosity and makes the code
easier to read.

Secondly, in newer code we use 'priv' (rather than privdata) and 'plat'
for private data but this is not consistent. In many places 'platdata' is
used for platform data. We can't really use 'pdata' since that is
ambiguous (it could mean private or platform data). So this series
standardises on 'plat'.

Finally, ofdata_to_platdata is shortened to of_to_plat since the 'data'
bits are redundant.

These changes apply tree-wide.
END

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I79a723ce639e0c0740d239dde73ce2d3d3528e09